[week3]카드 UI
