+++ week3/card_flip/src/styles/GlobalStyles.js
@@ -5,6 +5,8 @@ export const GlobalStyles = createGlobalStyle`
   ${reset}
   * {
     box-sizing : border-box;
+    max-width: 86rem;
+    margin: 4rem auto;
   }
   html {
     font-size: 62.5%;
